pull resources
